requirements.txt
Django==4.1.13
djangorestframework==3.14.0
python-decouple==3.8
python-dotenv==1.0.0
requests==2.28.1
whitenoise==6.9.0
Pillow==9.2.0
numpy==1.24.3
pandas==2.0.1
scikit-learn==1.6.1
scipy==1.10.1
plotly==5.14.1
django-cors-headers==4.3.1
django-crispy-forms==1.14.0
django-filter==22.1
django-taggit==3.1.0
django-reversion==5.0.4
django-bootstrap4==22.3
django-autocomplete-light==3.9.4
django-session-timeout==0.1.0
google-generativeai
razorpay
python-bidi==0.4.2
pytz==2022.6
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.12
PyPDF2==2.11.2
weasyprint==60.2
pytesseract==0.3.10
easy-thumbnails==2.8.3
soupsieve==2.3.2.post1
beautifulsoup4==4.11.1
Werkzeug==2.2.2
setuptools>=65.0.0
wheel==0.42.0
Django==4.1.13
djangorestframework==3.14.0
python-decouple==3.8
python-dotenv==1.0.0
requests==2.28.1
whitenoise==6.9.0
Pillow==9.2.0
numpy==1.24.3
pandas==2.0.1
scikit-learn==1.6.1
scipy==1.10.1
plotly==5.14.1
django-cors-headers==4.3.1
django-crispy-forms==1.14.0
django-filter==22.1
django-taggit==3.1.0
django-reversion==5.0.4
django-bootstrap4==22.3
django-autocomplete-light==3.9.4
django-session-timeout==0.1.0
google-generativeai
razorpay
python-bidi==0.4.2
pytz==2022.6
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.12
PyPDF2==2.11.2
weasyprint==60.2
pytesseract==0.3.10
easy-thumbnails==2.8.3
soupsieve==2.3.2.post1
beautifulsoup4==4.11.1
Werkzeug==2.2.2
